requirements.txt
python-telegram-bot
python-dotenv
paramiko
asyncssh
pathlib